feat: TimeComponent 컴포넌트 분리 작업 간 이슈 해결
